Rename PatientState to PatientCondition
