+++ inform-db/src/main/java/uk/ac/ucl/rits/inform/informdb/conditions/package-info.java
@@ -1,4 +1,4 @@
 /**
  * States that can affect a patient (e.g. diagnoses and infections).
  */
-package uk.ac.ucl.rits.inform.informdb.state;
+package uk.ac.ucl.rits.inform.informdb.conditions;
